A cleanup of the core files into separate files
