+++ packages/inferno/src/index.js
@@ -1,7 +1,7 @@
 import {
-	createVNode
-} from '../../../build/core/shapes';
-import cloneVNode from '../../../build/factories/cloneVNode';
+	createVNode,
+	cloneVNode
+} from '../../../build/core/VNodes';
 import { warning, NO_OP, isBrowser, EMPTY_OBJ } from '../../../build/shared';
 import { render, findDOMNode, createRenderer, enableFindDOMNode } from '../../../build/DOM/rendering';
 import { disableRecycling } from '../../../build/DOM/recycling';
