+++ src/component/es2015.ts
@@ -12,7 +12,7 @@ import {
 	Props,
 	VNode,
 	VNodeFlags
-} from '../core/shapes';
+} from '../core/structures';
 
 import Lifecycle from './../DOM/lifecycle';
