+++ src/core/VNodes.ts
@@ -1,20 +1,53 @@
 import {
-	Props,
-	InfernoChildren,
-	VNode,
-	VNodeFlags,
-	createVNode,
-	isVNode,
-} from '../core/shapes';
-import {
 	isArray,
-	isNull,
+	isInvalid,
 	isNullOrUndef,
 	isUndefined,
-	isInvalid
+	isNull,
+	isStatefulComponent
 } from '../shared';
+import {
+	normalize
+} from './normalization';
+import {
+	VNodeFlags,
+	VNode,
+	InfernoChildren,
+	Props,
+	Key,
+	Ref
+} from './structures';
 
-export default function cloneVNode(vNodeToClone: VNode, props?: Props, ..._children: InfernoChildren[]): VNode {
+export function createVNode(
+	flags: VNodeFlags,
+	type?,
+	props?: Props,
+	children?: InfernoChildren,
+	events?,
+	key?: Key,
+	ref?: Ref,
+	noNormalise?: boolean
+): VNode {
+	if (flags & VNodeFlags.ComponentUnknown) {
+		flags = isStatefulComponent(type) ? VNodeFlags.ComponentClass : VNodeFlags.ComponentFunction;
+	}
+	const vNode: VNode = {
+		children: isUndefined(children) ? null : children,
+		dom: null,
+		events: events || null,
+		flags: flags || 0,
+		key: key === undefined ? null : key,
+		props: props || null,
+		ref: ref || null,
+		type
+	};
+	if (!noNormalise) {
+		normalize(vNode);
+	}
+	return vNode;
+}
+
+export function cloneVNode(vNodeToClone: VNode, props?: Props, ..._children: InfernoChildren[]): VNode {
 	let children: any = _children;
 
 	if (_children.length > 0 && !isNull(_children[0])) {
@@ -102,4 +135,16 @@ export default function cloneVNode(vNodeToClone: VNode, props?: Props, ..._child
 	}
 	newVNode.dom = null;
 	return newVNode;
+}
+
+export function createVoidVNode(): VNode {
+	return createVNode(VNodeFlags.Void);
+}
+
+export function createTextVNode(text: string | number): VNode {
+	return createVNode(VNodeFlags.Text, null, null, text);
+}
+
+export function isVNode(o: VNode): boolean {
+	return !!o.flags;
 }
