+++ src/core/structures.ts
@@ -1,0 +1,57 @@
+export interface Styles {
+	[key: string]: number | string;
+}
+
+export interface IProps {
+	[index: string]: any;
+}
+export interface VType {
+	flags: VNodeFlags;
+}
+
+export type InfernoInput = VNode | VNode[] | null | string | string[] | number | number[];
+
+export const enum VNodeFlags {
+	Text = 1,
+	HtmlElement = 1 << 1,
+
+	ComponentClass = 1 << 2,
+	ComponentFunction = 1 << 3,
+	ComponentUnknown = 1 << 4,
+
+	HasKeyedChildren = 1 << 5,
+	HasNonKeyedChildren = 1 << 6,
+
+	SvgElement = 1 << 7,
+	MediaElement = 1 << 8,
+	InputElement = 1 << 9,
+	TextareaElement = 1 << 10,
+	SelectElement = 1 << 11,
+	Void = 1 << 12,
+	Element = HtmlElement | SvgElement | MediaElement | InputElement | TextareaElement | SelectElement,
+	Component = ComponentFunction | ComponentClass | ComponentUnknown
+}
+
+export type Key = string | number | null;
+export type Ref = Function | null;
+export type InfernoChildren = string | number | VNode | Array<string | VNode> | null;
+export type Type = string | Function | null;
+
+export interface Props {
+	children?: InfernoChildren;
+	ref?: Ref;
+	key?: Key;
+	events?: Object | null;
+}
+
+export interface VNode {
+	children: InfernoChildren;
+	dom: Node | null;
+	events: Object | null;
+	flags: VNodeFlags;
+	key: Key;
+	props: Props | null;
+	ref: Ref;
+	type: Type;
+	parentVNode?: VNode;
+}
