+++ src/redux/connect.ts
@@ -6,7 +6,7 @@ import {
 } from './utils';
 
 import Component from 'inferno-component';
-import { IProps } from '../core/shapes';
+import { IProps } from '../core/structures';
 import createElement from 'inferno-create-element';
 import hoistStatics from 'hoist-non-inferno-statics';
 import { isPlainObject } from './helpers';
